+++ src/api/main.go
@@ -3,12 +3,13 @@ package main
 import (
 	"github.com/marcosgfrites/REST-API_Golang/src/databases"
 	"github.com/marcosgfrites/REST-API_Golang/src/handlers"
+	"github.com/marcosgfrites/REST-API_Golang/src/utils/common"
 	"log"
 )
 
 func main() {
 	if databases.ConnectionCheck() == 0 {
-		log.Fatal("Database connection failed")
+		log.Fatal(common.DatabaseConnectionFailed)
 		return
 	}