+++ src/handlers/handlers.go
@@ -1,7 +1,10 @@
 package handlers
 
 import (
+	"fmt"
 	"github.com/gorilla/mux"
+	"github.com/marcosgfrites/REST-API_Golang/src/middlewares"
+	"github.com/marcosgfrites/REST-API_Golang/src/routers"
 	"github.com/rs/cors"
 	"log"
 	"net/http"
@@ -12,6 +15,9 @@ import (
 func Handlers() {
 	router := mux.NewRouter()
 
+	// routes
+	router.HandleFunc("/sign-up", middlewares.CheckDatabase(routers.Register)).Methods("POST")
+
 	port := os.Getenv("PORT")
 	if port == "" {
 		port = "8080"
@@ -20,5 +26,5 @@ func Handlers() {
 	//TODO: this permission will be changed. Now it allows all access
 	handler := cors.AllowAll().Handler(router)
 
-	log.Fatal(http.ListenAndServe(":"+port, handler))
+	log.Fatal(http.ListenAndServe(fmt.Sprintf(":%s", port), handler))
 }