+++ CHANGELOG.md
@@ -2,7 +2,14 @@
 
 ## [Unreleased]
 
+### X.X.X | dd-MM-yyyy
+- add feature to register new users
+- add routers
+- add user model
+- add middlewares
+
 ## [Released]
 
 ### 0.0.1 | 08-11-2022
 - add database connection
+- add handlers